split argumentScoreHandler
